Add production deployment scripts

+++ app/controllers/admin/cms_assets_controller.rb
@@ -3,6 +3,7 @@ module Admin
     def index
       @asset = CmsAsset.new
       @assets = CmsAsset.with_attached_file.order(:key)
+      @in_use_asset_keys = CmsAsset.in_use_keys
     end
 
     def create
@@ -12,6 +13,7 @@ module Admin
         redirect_to admin_cms_assets_path, notice: "Asset uploaded."
       else
         @assets = CmsAsset.with_attached_file.order(:key)
+        @in_use_asset_keys = CmsAsset.in_use_keys
         flash.now[:alert] = "Asset could not be uploaded."
         render :index, status: :unprocessable_entity
       end
@@ -19,6 +21,12 @@ module Admin
 
     def destroy
       asset = CmsAsset.find(params[:id])
+
+      if CmsAsset.in_use_keys.include?(asset.key)
+        redirect_to admin_cms_assets_path, alert: "Asset \"#{asset.key}\" is in use by CMS content or protected seed assets and was not deleted."
+        return
+      end
+
       asset.destroy
       redirect_to admin_cms_assets_path, notice: "Asset removed."
     end
